User prompting, refresh upload, add to latest transaction (#15)
This PR provides functionality for the MVP version of the app, which uploads a receipt and attaches it to the latest transaction the customer has. 
It also allows users to pull to refresh their transactions and rewards (next PR) screen.
If there has been a new transaction added after the refresh, the user will be prompted to upload a receipt.