Change to drawer navigation (#20)
Drawer navigation. 

We are no longer using the bottom tab navigation as of week 14 (from the Figma), so a drawer better serves our purpose. This PR revamps our navigation system and now maps all the routes in a drawer navigator instead of the tab navigator we had before

## Work done
Small frontend things. Added a green banner on the bottom of StoresScreen that takes you to RewardsScreen (homescreen in the code right now). Others include adding a hamburger that opens a drawer menu on each page, removing headers on all the pages, and a few small changes to the title text in announcements and history.